Try mixed Precision

+++ BA_ZHAW/models/beta_vanilla/train_beta_vanilla.py
@@ -244,8 +244,8 @@ def main():
         max_epochs=EPOCHS,
         logger=[wandb_logger, tensorboard_logger],
         callbacks=[model_checkpoint, early_stopping, lr_monitor, swa],  
-        accelerator="gpu",
-    )
+        accelerator="gpu", precision=16
+    ) # Mixed Precision Training verwenden mit precision=16 (Verringerung GPU Speicher)
 
     trainer.fit(model, train_dataloaders=train_dataloader, val_dataloaders=val_dataloader)
     best_model_path = model_checkpoint.best_model_path
